Split the info and announcements into separate tabs
